Add initial feed
The MCU board now lets the printer pull in the carrier until the new
paper in sensor is triggered for the first time and then starts the
print like before. The new "initial feed" provides an alignment of the
carrier which is exactly the same every time. Of course, the carrier now
needs two markings.